pyproject.toml
[project]
name = "avatar"
version = "0.1.0"
description = "Real-time AI conversational agent for Unreal MetaHuman control via Pipecat 0.95"
authors = [
    {name = "Admin Noosphere", email = "admin@noosphere.dev"}
]
readme = "README.md"
requires-python = ">=3.10,<3.13"
license = {text = "MIT"}
keywords = ["pipecat", "metahuman", "unreal", "ai", "ndi", "audio2face"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "pipecat-ai>=0.0.95",
    "websockets>=12.0",
    "aiohttp>=3.9.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "loguru>=0.7.0",
]

[project.optional-dependencies]
ndi = [
    "pygobject>=3.46.0",
]
dev = [
    "black>=23.12.0",
    "ruff>=0.1.9",
    "mypy>=1.8.0",
    "pytest>=7.4.3",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.23.2",
    "pre-commit>=3.6.0",
    "pytest-mock>=3.12.0",
]

[project.urls]
Homepage = "https://github.com/admin-noosphere/avatar"
Documentation = "https://github.com/admin-noosphere/avatar/blob/main/docs"
Repository = "https://github.com/admin-noosphere/avatar"
Issues = "https://github.com/admin-noosphere/avatar/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/avatar"]

# ============================================================================
# BLACK - Code Formatting
# ============================================================================
[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

# ============================================================================
# RUFF - Linting
# ============================================================================
[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "C",   # flake8-comprehensions
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by Black)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex (we handle this case by case)
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # Allow unused imports in __init__.py
"tests/*" = ["E501"]      # Allow long lines in tests

[tool.ruff.lint.isort]
known-first-party = ["avatar"]

# ============================================================================
# MYPY - Type Checking
# ============================================================================
[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_optional = true
strict_equality = true

# Allow missing imports for external libs
[[tool.mypy.overrides]]
module = [
    "websockets.*",
    "gi.*",
    "gi.repository.*",
    "pipecat.*",
]
ignore_missing_imports = true

# ============================================================================
# PYTEST - Testing
# ============================================================================
[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-ra",
    "--strict-markers",
    "--strict-config",
    "--cov=src/avatar",
    "--cov-report=term-missing:skip-covered",
    "--cov-report=html",
    "--cov-report=xml",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]
asyncio_mode = "auto"

# ============================================================================
# COVERAGE
# ============================================================================
[tool.coverage.run]
source = ["src"]
branch = true
omit = [
    "*/tests/*",
    "*/__init__.py",
    "*/site-packages/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
